extract new module
