+++ exif_edit/formats.py
@@ -136,25 +136,3 @@ class DegreeFormatFactory:
             return DecimalFormat(match.group(1))
 
         raise ValueError("unknown format to parse degree")
-
-class Coordinate:
-    """
-    Represent a coordinate on the globe.
-    """
-    def __init__(self, latitude, longitude, lat_ref = 'N', lon_ref = 'E'):
-        if not isinstance(latitude, Format):
-            latitude = DegreeFormatFactory.create(latitude)
-        self.latitude = latitude
-        self.lat_ref = lat_ref
-
-        if not isinstance(longitude, Format):
-            longitude = DegreeFormatFactory.create(longitude)
-        self.longitude = longitude
-        self.lon_ref = lon_ref
-
-    def decimal(self):
-        lat_dec = abs(self.latitude.as_float())
-        lon_dec = abs(self.longitude.as_float())
-        lat = lat_dec if self.lat_ref == 'N' else -1 * lat_dec
-        lon = lon_dec if self.lon_ref == 'E' else -1 * lon_dec
-        return lat, lon
